+++ workers/api/package.json
@@ -3,13 +3,13 @@
 	"version": "0.0.0",
 	"type": "module",
 	"scripts": {
-		"clean": "npx -y rimraf@latest ./pqc/container/dist",
+		"clean": "npx -y rimraf@latest ./container/pqc/dist",
 		"start": "wrangler dev --live-reload --persist-to ../.wrangler/state",
 		"build:types:cf": "wrangler types",
-		"build:pqc": "tsc --project pqc/container/tsconfig.json",
+		"build:pqc": "tsc --project container/pqc/tsconfig.json",
 		"build:static": "tsx staticApi.mts",
 		"base": "node --disallow-code-generation-from-strings --enable-source-maps --experimental-websocket --tls-min-v1.3 --trace-warnings --use-largepages=silent",
-		"start:pqc": "npm run base -- ./pqc/container/dist/index.mjs",
+		"start:pqc": "npm run base -- ./container/pqc/dist/index.mjs",
 		"wrangler": "npx --yes https://prerelease-registry.devprod.cloudflare.dev/workers-sdk/runs/14180338760/npm-package-wrangler-8740",
 		"publish:static": "tsx apiGateway.mts"
 	},
