+++ workers/api/src/v0/encrypt.mts
@@ -9,9 +9,9 @@ import { Buffer } from 'node:buffer';
 import { createCipheriv, createHash, createHmac, createSecretKey, hkdf, timingSafeEqual, type CipherKey } from 'node:crypto';
 import { promisify } from 'node:util';
 import type { ContextVariables, EnvVars } from '~/types.mjs';
+import type { routes as containerRoutes } from '~container/pqc/src/index.mjs';
 import { datakeys, keyrings } from '~db/tenant';
 import type { PqcContainerSidecar } from '~do/pqc.mjs';
-import type { routes as containerRoutes } from '~pqc/container/src/index.mjs';
 import { BitwardenHelper } from '~shared/helpers/bitwarden.mjs';
 import { cipherText0, type SecretNote } from '~shared/types/bw/index.mjs';
 import { EncryptionAlgorithms, KeyAlgorithms } from '~shared/types/crypto/index.mjs';
