+++ workers/api/tsconfig.json
@@ -5,6 +5,7 @@
 		"paths": {
 			"~/*": ["./src/*"],
 			"~db/*": ["../../db/*"],
+			"~container/*": ["./container/*"],
 			"~do/*": ["./do/*"],
 			"~shared/*": ["../../shared/*"],
 			"~pqc/*": ["./pqc/*"],
@@ -13,5 +14,5 @@
 		"types": ["./worker-configuration.d.ts", "node"]
 	},
 	"files": ["./eslint.config.mjs", "prettier.config.mjs", "staticApi.mjs", "apiGateway.mts"],
-	"include": ["../../db", "./do", "./wf", "src"]
+	"include": ["../../db", "./container", "./do", "./wf", "src"]
 }
